+++ CHANGELOG.md
@@ -66,7 +66,8 @@
   * Changed the `akamai_property_activation` resource that modifying `contact` field without changing `version`, `network` or `property_id` no longer shows diff ([I#624](https://github.com/akamai/terraform-provider-akamai/issues/624)).
   * Changed the `akamai_property_include_activation` resource that modifying `notify_emails` field without changing `version` no longer shows diff.
 
-
+* PAPI
+  * Changed validation of edgehostnames domain prefixes in `resource_akamai_edge_hostname` to create context.